requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Faker==23.0.0
pytest==8.0.0
requests==2.31.0

